% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing_functions.R
\name{compile_dictionary}
\alias{compile_dictionary}
\title{Compile a data dictionary for a data frame}
\usage{
compile_dictionary(dataset, name = "Data_Dictionary.csv", path = "./")
}
\arguments{
\item{dataset}{The dataframe to generate the dictionary}

\item{name}{The filename for the dictionary}

\item{path}{The filepath to locate the dictionary}
}
\value{
dict.temp The dictionary dataframe (Note: Not returned, the function will save the csv dictionary to the desired path)
}
\description{
Compiles a data dictionary for a data frame input
}
\examples{
my.dictionary <- websteR::compile_dictionary(iris)

}

