% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing_functions.R
\name{enforce_dictionary}
\alias{enforce_dictionary}
\title{Apply data dictionary to dataset}
\usage{
enforce_dictionary(data, dictionary.path, enforce.keep = TRUE)
}
\arguments{
\item{data}{Dataset to be processed}

\item{dictionary.path}{Filepath to the data dictionary to be enforced}

\item{enforce.keep}{When enforce.keep = TRUE (Default), filters out columns with a "keep" value of FALSE in the dictionary}
}
\value{
data Data frame with dictionary rules applied
}
\description{
Applies typing and filtering rules from a data dictionary to a dataset
}

